created graph layout
