+++ pages/graph.tsx
@@ -1,0 +1,5 @@
+import GraphLayout from "../layouts/graph";
+
+export default function Graph() {
+  return <GraphLayout />;
+}
